NOT_FOR_MERGE Test build

+++ tensorflow/tools/ci_build/builds/pip_new.sh
@@ -326,8 +326,11 @@ if [[ -z "$PYTHON_BIN_PATH" ]]; then
 fi
 
 ${PYTHON_BIN_PATH} -m pip install tb-nightly
-${PYTHON_BIN_PATH} -m pip uninstall -y protobuf
-${PYTHON_BIN_PATH} -m pip install "protobuf < 4"
+if [[ "x${PY_MAJOR_MINOR_VER}x" == "x3.8x" ]]; then
+  ls -al /usr/local/lib/python3.8/dist-packages/google/protobuf/
+  ${PYTHON_BIN_PATH} -m pip uninstall -y protobuf
+  ${PYTHON_BIN_PATH} -m pip install "protobuf < 4"
+fi
 
 # Bazel build the file.
 PIP_BUILD_TARGET="//tensorflow/tools/pip_package:build_pip_package"
